skills/yuriyanzexuan/mm-output/requirements.txt
torch
git+https://github.com/Hadlay-Zhang/marker.git
marker-pdf
Pillow
markdown
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Multi-modal output dependencies
playwright>=1.40.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
